requirements.txt
shapely==2.0.2
scikit-image==0.22.0
imgaug==0.4.0
pyclipper==1.3.0.post5
lmdb==1.4.1
tqdm==4.66.1
numpy==1.24.3
visualdl==2.5.3
rapidfuzz==3.6.1
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
opencv-python-headless==4.9.0.80
cython==3.0.8
lxml==5.1.0
premailer==3.10.0
openpyxl==3.1.2
attrdict==2.0.1
PyMuPDF<=1.21.0
Pillow>=10.0.0
pyyaml==6.0.1
paddlepaddle==2.6.0
PyQt5==5.15.10
ultralytics==8.0.200
unidecode==1.3.8
paddleocr==2.7.0.3
onnxruntime==1.16.3
pandas
torch
xlwings==0.30.15
psutil